Smart Attendance/backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
Flask-SocketIO==5.3.6